zlib is now optional.
`USING_ZLIB` must be defined to enable parsing of compressed record
fields, otherwised they are skipped.

+++ src/generic/file.h
@@ -29,7 +29,6 @@
 #include <vector>
 #include <stdint.h>
 
-#include <zlib.h>
 #include <boost/crc.hpp>
 
 namespace espm {
@@ -62,7 +61,7 @@ namespace espm {
             }
 
             input.seekg(0, input.end);
-            unsigned int length = input.tellg();
+            size_t length = input.tellg();
             input.seekg(0, input.beg);
 
             //Allocate memory for file contents.
